refactor: Java 5 issues (#2111)
* refactor: using RoundingMode enum constant

* refactor: using enhanced for loop

* refactor: using enhanced 'for' loop instead of 'while' loop

* refactor: using 'String.contains' instead of 'String.indexOf()' expression

* refactor: removing unnecessary boxing

* refactor: using 'Integer.parseInt' instead of 'Integer.valueOf'

* refactor: removing unnecessary unboxing

* refactor: using 'StringBuilder' instead of 'StringBuffer'

---------

+++ src/main/java/module/teamAnalyzer/ui/renderer/PlayerPositionTableCellRenderer.java
@@ -39,7 +39,7 @@ public class PlayerPositionTableCellRenderer extends DefaultTableCellRenderer {
         if (value instanceof Integer) {
             super.getTableCellRendererComponent(table, null, isSelected, hasFocus, row, col);
 
-            int pos = ((Integer) value).intValue();
+            int pos = (Integer) value;
 
             Icon icon;
